Added chrome dev tools support
