+++ src/index.prod.js
@@ -1,0 +1,22 @@
+import 'whatwg-fetch';
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { Provider } from 'react-redux';
+import Router from 'react-router/BrowserRouter';
+
+import App from './App';
+import './index.css';
+import configureStore from './configureStore.js';
+
+const store = configureStore();
+
+ReactDOM.render(
+  (
+    <Provider store={store}>
+      <Router>
+        <App />
+      </Router>
+    </Provider>
+  ),
+  document.getElementById('root')
+);
